@@ -16,27 +16,32 @@
     "php": "^8.1",
     "ext-ctype": "*",
     "ext-iconv": "*",
-    "symfony/asset": "^6.4",
-    "symfony/config": "^6.4",
-    "symfony/dependency-injection": "^6.4",
-    "symfony/dotenv": "^6.4",
-    "symfony/framework-bundle": "^6.4",
-    "symfony/http-kernel": "^6.4",
-    "symfony/monolog-bundle": "^3.8",
-    "symfony/orm-pack": "^2.0",
-    "symfony/routing": "^6.4",
-    "symfony/security-bundle": "^6.4",
-    "symfony/translation": "^6.4",
-    "symfony/twig-bundle": "^6.4",
-    "symfony/validator": "^6.4",
-    "symfony/yaml": "^6.4",
-    "symfony/form": "^6.4",
+    "symfony/asset": "^7.0",
+    "symfony/config": "^7.0",
+    "symfony/dependency-injection": "^7.0",
+    "symfony/dotenv": "^7.0",
+    "symfony/framework-bundle": "^7.0",
+    "symfony/http-kernel": "^7.0",
+    "symfony/monolog-bundle": "^3.10",
+    "symfony/orm-pack": "^2.3",
+    "symfony/routing": "^7.0",
+    "symfony/security-bundle": "^7.0",
+    "symfony/translation": "^7.0",
+    "symfony/twig-bundle": "^7.0",
+    "symfony/validator": "^7.0",
+    "symfony/yaml": "^7.0",
+    "symfony/form": "^7.0",
     "doctrine/annotations": "^1.0",
-    "sensio/framework-extra-bundle": "^6.2",
-    "symfony/mailer": "^6.4"
+    "symfony/mailer": "^7.0",
+    "symfony/runtime": "^7.3"
   },
   "require-dev": {
-    "symfony/maker-bundle": "^1.41",
-    "symfony/phpunit-bridge": "^6.4"
+    "symfony/maker-bundle": "^1.63",
+    "symfony/phpunit-bridge": "^7.0"
+  },
+  "config": {
+    "allow-plugins": {
+      "symfony/runtime": true
+    }
   }
 }
